pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "An automated agent for job and freelance opportunity search"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "selenium>=4.15.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "langchain>=0.0.300",
    "pandas>=2.1.0",
    "jinja2>=3.1.2",
    "schedule>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"
